ev-window: Fixed a problem where the find-bar would be displayed while in the recent-view
The find-bar had been moved by KaL while implementing the sidebar. On
rebasing the bookshelf branch, there was a tiny issue - if a user switched
to the bookshelf/recent view while in search mode in a document, the
find-bar would remain visible in the bookshelf/recent view. Just addded a
line to hide the find-bar when switching to recent view.